refactor(math): relocate screen-scale disc sizing to @carma-commons/math
The pure screen-targeted sizing maths are not gizmo-specific (the point-query
disc and other known-scale reference objects use them too), so move them to the
lowest-level math commons with implementation-neutral names:

- resolveGizmoDiscWorldRadius -> resolveWorldSizeForScreenTarget
- shouldRestepGizmoDisc       -> shouldRestepScreenScale
- snapWorldRadiusToNiceStep   -> snapToNiceStep
- GIZMO_DISC_STEP_FACTOR      -> SCREEN_SCALE_STEP_FACTOR
- computeGizmoDiscSegments    -> computeCircleSegments

gizmo/core keeps only GIZMO_DISC_RESIZE_TRIGGERS (gizmo UI semantics). Importers
(gizmo/cesium, point-query indicator controller) now pull the generic helpers
from @carma-commons/math.
