+++ libraries/mapping/gizmo/core/src/index.ts
@@ -7,12 +7,5 @@ export * from "./lib/svgProjection";
 export * from "./lib/projectedMoveGizmoView";
 export {
   GIZMO_DISC_RESIZE_TRIGGERS,
-  GIZMO_DISC_STEP_FACTOR,
-  computeGizmoDiscSegments,
-  resolveGizmoDiscWorldRadius,
-  shouldRestepGizmoDisc,
-  snapWorldRadiusToNiceStep,
   type GizmoDiscResizeTrigger,
-  type GizmoDiscSegmentOptions,
-  type ResolveGizmoDiscWorldRadiusInput,
 } from "./lib/gizmoDiscSizing";
